Seeders added to for Stripe integration
For the stripe integration to work seeders were needed because the data of the product should be mentioned in the stripe dashboard as well. So the fix for this was to get all the products from the API and plant them into the stripe dashboard which would give us a price_id and we can use it to successfully append the stripe checkout process. The price_id will be seeded into the product.